oz-Django/pyproject.toml
[tool.poetry]
name = "oz-django"
version = "0.1.0"
description = "oz python django project"
authors = ["hotbari <ghkahr01@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
